[driver] Move BlockAllocator out of :architecture:heap
