+++ src/modm/driver/storage/block_allocator.lb
@@ -1,0 +1,21 @@
+# Copyright (c) 2020, Niklas Hauser
+#
+# This file is part of the modm project.
+#
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+# -----------------------------------------------------------------------------
+
+def init(module):
+    module.name = ":driver:block.allocator"
+    module.description = "Block Allocator"
+
+def prepare(module, options):
+    module.depends(":architecture", ":math:utils")
+    return True
+
+def build(env):
+    env.outbasepath = "modm/src/modm/driver/storage"
+    env.copy("block_allocator.hpp")
+    env.copy("block_allocator_impl.hpp")
